Added support for editing surveys
