+++ donut/modules/voting/templates/edit.html
@@ -96,7 +96,7 @@
 
 {% block page %}
   <div class="col-md-8 col-md-offset-2" style="float: none; padding-bottom: 25px">
-    <h1>Editting survey questions for {{ survey['title'] }}</h1>
+    <h1>Editing survey questions for {{ survey['title'] }}</h1>
     {% if survey['description'] %}
       <h3>{{ survey['description'] }}</h3>
     {% endif %}
